Volume: Factor out BufferSource, BufferedVolume and RAIVolume
